Add default cases to switch statements
When compiling with -Wall, compiler warnings will be spawned from
client code.  This forces clients to disable warnings, which is
a bad thing.

This patch will make the warnings go away: instead of having a switch
_and_ a return statement for when no case was matched, now the
return statement is moved into the switches default case.

I believe this is equivalent, but improves the symmetry of the code.